Fixing a problem with BillAgreementUpdate
As the above-mentioned call does require a request structure that
does not fit the default, it must be handled differently.

Problem occurs when using the Merchant SDK [1]. Solution seems to
have been given by PayPal support once[2] but was obviously not
handed to the ruby team.

Symptom would be that the call returns an error 10006 ("Version is
not supported") as the server can't find any request content.

[1] https://github.com/paypal/merchant-sdk-ruby/
[2] https://go.developer.ebay.com/developers/paypal/forums/soap/billagreementupdate-why-giving-me-version-not-supported-error